kcov uses cki gcov kernels

See merge request cki-project/kernel-tests!904

+++ kcov/prepare/runtest.sh
@@ -24,29 +24,22 @@ set -x
 
 TEST="/kcov/prepare"
 
-RHEL=$(sed -ne's/^Red Hat.*release \([0-9]\.[0-9]\).*$/\1/p' /etc/redhat-release)
-MAJOR=$(echo $RHEL | cut -c1)
-MINOR=$(echo $RHEL | cut -c3)
-VERSION=${RHEL_KRNL[$MAJOR$MINOR]}
-
 setup()
 {
-    KN=kernel
-    GCOVKERNEL=$KN-$VERSION.gcov
-    GCOVCOREKERNEL=$KN-core-$VERSION.gcov
-    GCOVMODULES=$KN-modules-$VERSION.gcov
-    GCOVDA=$KN-gcov-$VERSION.gcov
-
     # prepare the kernel for coverage collection
     log "Red Hat release: $(cat /etc/redhat-release)"
     log "Running on $(arch), kernel $(uname -r)"
 
     # setting configs
     log "setting configs"
-        if [ -n "$KDIR" ]; then
+    if [ -n "$KDIR" ]; then
         KCOV_KDIR=$KDIR
-        fi
+    fi
 
+    if [ ! -e /sys/kernel/debug/gcov ]; then
+        fail prepare "kernel doesn't seem to support gcov";
+        exit
+    fi
 
     log "write config"
     echo "KDIR=$KCOV_KDIR" > $KCOV_CONF
@@ -56,30 +49,15 @@ setup()
     log "submit config"
     rhts-submit-log -l $KCOV_CONF
 
-   #REPO is defined on include.sh
-   repo_url="${REPO}/${VERSION}.gcov/`arch`/"
-
-   log "Configuring kcov kernel repo ($repo_url)"
-   cat << EOF > /etc/yum.repos.d/kernel-gcov.repo
-[kernel-gcov]
-name=kernel-gcov
-baseurl=${repo_url}
-enabled=1
-gpgcheck=0
-EOF
-
-    log "install kcov kernel"
-    dnf install -y $GCOVKERNEL $GCOVCOREKERNEL $GCOVMODULES
+    rpm -q kernel-gcov
     if [ $? -ne 0 ]; then
-        fail prepare "Cannot install the kcov kernel package.";
-        exit
-    fi
-
-    log "install gcov data files"
-    dnf install -y $GCOVDA
-    if [ $? -ne 0 ]; then
-        fail prepare "Cannot install the gcov data files package.";
-        exit
+        # Expects cki repo with kernel rpms to be configured
+        log "install gcov data files"
+        dnf install -y kernel-gcov
+        if [ $? -ne 0 ]; then
+            fail prepare "Cannot install the gcov data files package.";
+            exit
+        fi
     fi
 
     install_lcov
@@ -97,11 +75,7 @@ EOF
     # see Bug 1290759 for details
     echo "geninfo_gcov_all_blocks = 0" >> /etc/lcovrc
 
-    log "current boot kernel is: $(grubby --default-kernel)"
-    log "change boot kernel to gcov kernel"
-    grubby --set-default=$KCOV_IMG
-    log "new boot kernel is: $(grubby --default-kernel)"
-    log "reboot to gcov kernel"
+    log "reboot to make sure all the changes are in effect"
 
     touch ./kernel_installed
 
@@ -112,19 +86,12 @@ verify()
 {
     log "after reboot"
     log "current kernel is $(uname -r)"
-    uname -r | grep gcov
-    if [ $? -ne 0 ]; then
+    if [ ! -e /sys/kernel/debug/gcov ]; then
         fail "prapare" "not running on gcov kernel."
         fail
         exit
     fi
 
-    log "ok, going on the coverage testing..."
-
-    mount | grep debugfs
-    if [ $? -ne 0 ]; then
-        mount none /sys/kernel/debug -t debugfs
-    fi
     log "proc entries: $(ls /sys/kernel/debug/gcov)"
 
     pass
